Added run_query function that rolls back transaction if db error occurs
This function will issue a rollback when a database connection error occur during the execution of the passed query. By doing this sqlalchemy can replay the transaction when it manages to connect to the database again so that the previous statements will not be lost.

The implementation comes from here: https://stackoverflow.com/questions/29634488/better-approach-to-handling-sqlalchemy-disconnects

Additional changes:
* Use `tox-travis` to set up python environments
* Start using docker in travis
* Add toxiproxy and mysql container setup to Makefile
* Stop passing `worker_ctx_cls` argument to `ServiceContainer`